Move recording logic into separate file
